+++ packages/components/elements/floatBox/FloatBox.tsx
@@ -8,6 +8,8 @@ export type FloatBoxProps = {
   children?: React.ReactNode;
   side?: "bottom" | "left" | "right" | "top";
   sideOffset?: number;
+  align?: "start" | "center" | "end";
+  alignOffset?: number;
 };
 
 const FloatBox: React.FC<FloatBoxProps> = ({
@@ -15,28 +17,39 @@ const FloatBox: React.FC<FloatBoxProps> = ({
   open,
   side = "bottom",
   sideOffset = 40,
+  align = "center",
   ...props
 }) => {
-  let widthStyles = {
-    trigger: "var(--radix-popover-trigger-width)",
-    default: "auto",
-  };
-
-  let sideOffsetStyles = {
-    bottom: { top: sideOffset },
-    top: { bottom: sideOffset },
-    right: { left: sideOffset },
-    left: { right: sideOffset },
+  let stylesMap = {
+    bottom: {
+      start: { top: sideOffset, insetInlineStart: 0 },
+      center: { top: sideOffset },
+      end: { top: sideOffset, insetInlineEnd: 0 },
+    },
+    top: {
+      start: { bottom: sideOffset, insetInlineStart: 0 },
+      center: { bottom: sideOffset },
+      end: { bottom: sideOffset, insetInlineEnd: 0 },
+    },
+    right: {
+      start: { left: sideOffset, top: -5 },
+      center: { left: sideOffset },
+      end: { left: sideOffset, bottom: 0 },
+    },
+    left: {
+      start: { right: sideOffset, top: 0 },
+      center: { right: sideOffset },
+      end: { right: sideOffset, bottom: 0 },
+    },
   };
 
   return (
     <div
       className={cn(
         "data-[floatbox-state=closed]:hawa-invisible data-[floatbox-state=open]:hawa-visible hawa-absolute dark:dark-shadow hawa-z-50 hawa-rounded hawa-border hawa-bg-popover hawa-text-popover-foreground hawa-shadow-md hawa-outline-none data-[floatbox-state=open]:hawa-animate-in data-[floatbox-state=closed]:hawa-animate-out data-[floatbox-state=closed]:hawa-fade-out-0 data-[floatbox-state=open]:hawa-fade-in-0 data-[floatbox-state=closed]:hawa-zoom-out-95 data-[floatbox-state=open]:hawa-zoom-in-95 data-[side=bottom]:hawa-slide-in-from-top-2 data-[side=left]:hawa-slide-in-from-right-2 data-[side=right]:hawa-slide-in-from-left-2 data-[side=top]:hawa-slide-in-from-bottom-2",
-        sideOffsetStyles[side],
         className,
       )}
-      style={{ ...sideOffsetStyles[side] }}
+      style={{ ...stylesMap[side][align] }}
       data-side={side}
       data-floatbox-state={open ? "open" : "closed"}
     >