+++ packages/components/hooks/index.ts
@@ -14,3 +14,4 @@ export * from "./useMediaQuery";
 export * from "./useScrollPosition";
 export * from "./useTable";
 export * from "./useTabs";
+export * from "./useMeasureDirty";